refactor: #4 implement customer auth module
