--- src/main/kotlin/com/carava/carwash/auth/entity/UserType.kt
@@ -1,5 +1,0 @@
-package com.carava.carwash.auth.entity
-
-enum class UserType {
-    OWNER, CUSTOMER
-}
