+++ src/main/kotlin/com/carava/carwash/customer/dto/CheckUsernameResponseDto.kt
@@ -1,4 +1,4 @@
-package com.carava.carwash.auth.dto
+package com.carava.carwash.customer.dto
 
 data class CheckUsernameResponseDto (
     val isDuplicate: Boolean
