+++ src/main/kotlin/com/carava/carwash/customer/dto/SignInResponseDto.kt
@@ -1,4 +1,4 @@
-package com.carava.carwash.auth.dto
+package com.carava.carwash.customer.dto
 
 data class SignInResponseDto (
     val accessToken: String,
