+++ src/main/kotlin/com/carava/carwash/customer/dto/SignUpRequestDto.kt
@@ -1,4 +1,4 @@
-package com.carava.carwash.auth.dto
+package com.carava.carwash.customer.dto
 
 data class SignUpRequestDto (
     val email: String,
