+++ src/main/kotlin/com/carava/carwash/customer/repository/AuthRepository.kt
@@ -1,4 +1,12 @@
 package com.carava.carwash.customer.repository
 
-interface AuthRepository {
+import com.carava.carwash.customer.entity.Auth
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+import java.util.*
+
+@Repository
+interface AuthRepository : JpaRepository<Auth, Long> {
+    fun findByEmail(email: String): Optional<Auth>
+    fun existsByEmail(email: String): Boolean
 }
