+++ src/main/kotlin/com/carava/carwash/customer/service/CustomUserDetailsService.kt
@@ -1,18 +1,20 @@
-package com.carava.carwash.auth.service
+package com.carava.carwash.customer.service
 
-import com.carava.carwash.auth.repository.AuthRepository
+import com.carava.carwash.customer.repository.AuthRepository
+import com.carava.carwash.customer.security.CustomerUserDetails
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.core.userdetails.UserDetailsService
 import org.springframework.security.core.userdetails.UsernameNotFoundException
 import org.springframework.stereotype.Service
 
 @Service
-class CustomUserDetailsService(
+class CustomUserDetailsService (
     private val authRepository: AuthRepository
 ) : UserDetailsService {
-
     override fun loadUserByUsername(username: String): UserDetails {
-        return authRepository.findByEmail(username)
-            .orElseThrow { UsernameNotFoundException("사용자를 찾을 수 없습니다: $username") }
+        val auth = authRepository.findByEmail(username)
+            .orElseThrow { UsernameNotFoundException("사용자를 찾을 수 없습니다.") }
+
+        return CustomerUserDetails(auth)
     }
 }
