+++ src/main/kotlin/com/carava/carwash/global/config/JpaConfig.kt
@@ -1,0 +1,9 @@
+package com.carava.carwash.global.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing
+
+@Configuration
+@EnableJpaAuditing
+class JpaConfig {
+}
