+++ src/main/kotlin/com/carava/carwash/global/constants/UserType.kt
@@ -1,0 +1,6 @@
+package com.carava.carwash.global.constants
+
+enum class UserType {
+    CUSTOMER,
+    OWNER
+}
