+++ src/main/kotlin/com/carava/carwash/global/exception/EmailAlreadyExistsException.kt
@@ -1,3 +1,3 @@
-package com.carava.carwash.auth.error
+package com.carava.carwash.global.exception
 
 class EmailAlreadyExistsException(message: String) : RuntimeException(message)
